+++ lib/src/mobile_scanner_controller.dart
@@ -99,19 +99,21 @@ class MobileScannerController {
 
   bool isStarting = false;
 
-  bool? _hasTorch;
+  /// A notifier that provides availability of the Torch (Flash)
+  final ValueNotifier<bool?> hasTorchState = ValueNotifier(false);
 
   /// Returns whether the device has a torch.
   ///
   /// Throws an error if the controller is not initialized.
   bool get hasTorch {
-    if (_hasTorch == null) {
+    final hasTorch = hasTorchState.value;
+    if (hasTorch == null) {
       throw const MobileScannerException(
         errorCode: MobileScannerErrorCode.controllerUninitialized,
       );
     }
 
-    return _hasTorch!;
+    return hasTorch;
   }
 
   /// Set the starting arguments for the camera
@@ -123,6 +125,15 @@ class MobileScannerController {
     arguments['torch'] = torchEnabled;
     arguments['speed'] = detectionSpeed.index;
     arguments['timeout'] = detectionTimeoutMs;
+
+    /*    if (scanWindow != null) {
+      arguments['scanWindow'] = [
+        scanWindow!.left,
+        scanWindow!.top,
+        scanWindow!.right,
+        scanWindow!.bottom,
+      ];
+    } */
 
     if (formats != null) {
       if (Platform.isAndroid) {
@@ -210,8 +221,9 @@ class MobileScannerController {
       );
     }
 
-    _hasTorch = startResult['torchable'] as bool? ?? false;
-    if (_hasTorch! && torchEnabled) {
+    final hasTorch = startResult['torchable'] as bool? ?? false;
+    hasTorchState.value = hasTorch;
+    if (hasTorch && torchEnabled) {
       torchState.value = TorchState.on;
     }
 
@@ -223,7 +235,7 @@ class MobileScannerController {
               startResult['videoHeight'] as double? ?? 0,
             )
           : toSize(startResult['size'] as Map? ?? {}),
-      hasTorch: _hasTorch!,
+      hasTorch: hasTorch,
       textureId: kIsWeb ? null : startResult['textureId'] as int?,
       webId: kIsWeb ? startResult['ViewID'] as String? : null,
     );
@@ -244,7 +256,7 @@ class MobileScannerController {
   ///
   /// Throws if the controller was not initialized.
   Future<void> toggleTorch() async {
-    final hasTorch = _hasTorch;
+    final hasTorch = hasTorchState.value;
 
     if (hasTorch == null) {
       throw const MobileScannerException(
@@ -314,6 +326,8 @@ class MobileScannerController {
           BarcodeCapture(
             barcodes: parsed,
             image: event['image'] as Uint8List?,
+            width: event['width'] as double?,
+            height: event['height'] as double?,
           ),
         );
         break;
@@ -349,5 +363,11 @@ class MobileScannerController {
       default:
         throw UnimplementedError(name as String?);
     }
+  }
+
+  /// updates the native scanwindow
+  Future<void> updateScanWindow(Rect window) async {
+    final data = [window.left, window.top, window.right, window.bottom];
+    await _methodChannel.invokeMethod('updateScanWindow', {'rect': data});
   }
 }